// src/types/VehicleDetail.ts
export type VehicleDetail = {
  id: string;
  adUrl: string;
  breadcrumbs: { text: string; link: string }[];
  categoryName: string | null;
  className: string;
  condition: string;
  customerId: string;
  dealerId: string;
  dealerLogo: { url: string };
  dealerMapUrl: string;
  dealerName: string;
  dealerRepeatTag: string;
  features: { id: string; name: string }[];
  makeAnOfferPrice: string | null;
  makeDisplayName: string;
  modelDisplayName: string;
  phone: string;
  photos: { url: string }[];
  price: string;
  realmId: string;
  state: string;
  trimName: string;
  year: string;
  zip: string;
  formattedPhone: string;
  description: string;
  miles: string;
  location: string;
  locationDistance: string;
  details: { [key: string]: string };
  factorySpecs: string | null;
  localDelivery: string;
  emailSellerUrl: string;
  dealerLink: string;
  realmUrl: string;
  realmName: string;
  isSmsAvailable: boolean;
  isChatAvailable: boolean;
  isDealerWebsiteAvailable: boolean;
  video?: { thumbnail: string; url: string };
  vrUrl: string;
  floorPlanMediaId: string | null;
};